boh/ec2-user@ec2-34-201-140-36.compute-1.amazonaws.com/requirements.txt
alembic==1.13.2
aniso8601==9.0.1
astroid==3.3.3
Authlib==1.3.2
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.41.0
cryptography==43.0.1
dill==0.3.8
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.1.0
gunicorn==23.0.0
idna==3.10
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.0
mccabe==0.7.0
packaging==24.1
platformdirs==4.3.6
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.4.0
pylint==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
sib-api-v3-sdk==7.6.0
six==1.16.0
SQLAlchemy==2.0.35
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.2.3
watchdog==5.0.2
Werkzeug==3.0.4
